+++ credmark/cmf/engine/context.py
@@ -12,7 +12,7 @@ from credmark.cmf.engine.errors import ModelRunRequestError
 from credmark.cmf.engine.mocks import ModelMockException, ModelMockRunner
 from credmark.cmf.engine.model_api import ModelApi
 from credmark.cmf.engine.model_loader import ModelLoader
-from credmark.cmf.engine.web3_registry import Web3Registry
+from credmark.cmf.engine.web3.registry import Web3Registry
 from credmark.cmf.model import IncrementalModel, Model
 from credmark.cmf.model.context import ModelContext
 from credmark.cmf.model.errors import (