+++ credmark/cmf/engine/model_unittest.py
@@ -9,7 +9,7 @@ from credmark.cmf.engine.context import EngineModelContext
 from credmark.cmf.engine.mocks import ModelMockConfig, ModelMockRunner
 from credmark.cmf.engine.model_api import ModelApi
 from credmark.cmf.engine.model_loader import ModelLoader
-from credmark.cmf.engine.web3_registry import Web3Registry
+from credmark.cmf.engine.web3.registry import Web3Registry
 from credmark.cmf.model.context import ModelContext