+++ credmark/cmf/engine/web3/registry.py
@@ -5,7 +5,7 @@ from typing import Optional, Union
 
 from web3 import AsyncHTTPProvider, AsyncWeb3, HTTPProvider, Web3, WebsocketProvider
 
-from credmark.cmf.engine.web3_helper import add_chain_id_cache, add_chain_id_cache_async, inject_poa
+from credmark.cmf.engine.web3.helper import add_chain_id_cache, add_chain_id_cache_async, inject_poa
 from credmark.cmf.types.network import CREDMARK_PUBLIC_PROVIDERS, Network
 
 
@@ -92,19 +92,19 @@ class Web3Registry:
         if chain_to_provider_url:
             self.__chain_to_provider_url.update(chain_to_provider_url)
 
-    # pylint:disable=line-too-long
-    def web3_for_chain_id(self, chain_id: int, request_kwargs: Optional[dict] = None):
+    def provider_url_for_chain_id(self, chain_id: int):
         url = self.__chain_to_provider_url.get(str(chain_id))
         if url is None:
             raise Exception(
                 f'No web3 provider url for chain id {chain_id}. '
-                "In .env file or environment, set CREDMARK_WEB3_PROVIDERS as {'1':'https://web3-node-provider-url'}.")
-        return self.web3_for_provider_url(url, chain_id, request_kwargs)
+                'In .env file or environment, set CREDMARK_WEB3_PROVIDERS as '
+                f'{"{chain_id}":"https://web3-node-provider-url"}.')
+        return url
+
+    def web3_for_chain_id(self, chain_id: int, request_kwargs: Optional[dict] = None):
+        provider_url = self.provider_url_for_chain_id(chain_id)
+        return self.web3_for_provider_url(provider_url, chain_id, request_kwargs)
 
     def async_web3_for_chain_id(self, chain_id: int, request_kwargs: Optional[dict] = None):
-        url = self.__chain_to_provider_url.get(str(chain_id))
-        if url is None:
-            raise Exception(
-                f'No web3 provider url for chain id {chain_id}. '
-                "In .env file or environment, set CREDMARK_WEB3_PROVIDERS as {'1':'https://web3-node-provider-url'}.")
-        return self.async_web3_for_provider_url(url, chain_id, request_kwargs)
+        provider_url = self.provider_url_for_chain_id(chain_id)
+        return self.async_web3_for_provider_url(provider_url, chain_id, request_kwargs)