+++ tests/test_multicall.py
@@ -14,19 +14,18 @@ class TestModel(ModelTestCase):
     def test_try_aggregate(self):
         t = Token('USDC')
 
-        m = credmark.cmf.model.ModelContext.current_context().multicall
-        [symbol_result, decimals_result] = m.try_aggregate([t.functions.symbol(), t.functions.decimals()])
-
+        batch = credmark.cmf.model.ModelContext.current_context().web3_batch
+        [symbol_result, decimals_result] = batch.call([t.functions.symbol(), t.functions.decimals()])
         self.assertEqual(symbol_result.success, True)
         self.assertEqual(symbol_result.return_data_decoded, 'USDC')
 
         self.assertEqual(decimals_result.success, True)
         self.assertEqual(decimals_result.return_data_decoded, 6)
 
-        [symbol_result2, decimals_result2] = m.try_aggregate_unwrap([t.functions.symbol(), t.functions.decimals()])
+        [symbol_result2, decimals_result2] = batch.call([t.functions.symbol(), t.functions.decimals()], unwrap=True)
 
-        self.assertEqual(symbol_result2, symbol_result)
-        self.assertEqual(decimals_result2, decimals_result)
+        self.assertEqual(symbol_result2, symbol_result.return_data_decoded)
+        self.assertEqual(decimals_result2, decimals_result.return_data_decoded)
 
     @model_context(chain_id=1, block_number=17_000_000)
     def test_try_aggregate_same_function(self):
@@ -34,9 +33,9 @@ class TestModel(ModelTestCase):
         t1 = Token('USDC')
         t2 = Token('WETH')
 
-        m = credmark.cmf.model.ModelContext.current_context().multicall
-        results = m.try_aggregate_same_function(t1.functions.balanceOf(Address(address).checksum),
-                                                [t1.address.checksum, t2.address.checksum])
+        batch = credmark.cmf.model.ModelContext.current_context().web3_batch
+        results = batch.call_same_function(t1.functions.balanceOf(Address(address).checksum),
+                                           [t1.address.checksum, t2.address.checksum])
 
         b1 = results[0]
         b2 = results[1]
@@ -47,9 +46,10 @@ class TestModel(ModelTestCase):
         self.assertEqual(b2.success, True)
         self.assertEqual(b2.return_data_decoded, 0)
 
-        results2 = m.try_aggregate_same_function_unwrap(
+        results2 = batch.call_same_function(
             t1.functions.balanceOf(Address(address).checksum),
-            [t1.address.checksum, t2.address.checksum])
+            [t1.address.checksum, t2.address.checksum],
+            unwrap=True)
 
         self.assertEqual(results2[0], b1.return_data_decoded)
         self.assertEqual(results2[1], b2.return_data_decoded)
@@ -59,9 +59,9 @@ class TestModel(ModelTestCase):
         t1 = Token('USDC').as_erc20(True)
         t2 = Token('WETH').as_erc20(True)
 
-        m = credmark.cmf.model.ModelContext.current_context().multicall
-        [b1, b2] = m.try_aggregate_same_function(t1.functions.DECIMALS(),
-                                                 [t1.address.checksum, t2.address.checksum])
+        batch = credmark.cmf.model.ModelContext.current_context().web3_batch
+        [b1, b2] = batch.call_same_function(t1.functions.DECIMALS(),
+                                            [t1.address.checksum, t2.address.checksum])
 
         self.assertEqual(b1.success, False)
         self.assertEqual(b1.return_data_decoded, b"")